refactor(with-storage-sync): mv file
